moved static jsons to data folder
